Determining play url, now just before play instead of at list creation
